Appointments now write into an array so multiple can be loaded and displayed over a dynamic article set in the html. Location info is stored in a seperate array and referenced as per the appointment.
Messages now also write into an array. The thread array stores the patient and doctor info and the threadMessages array stores an array of messages that make up a thread. The navigation for the 3  dimensional array can be found in its declaration at the top. The display also has logic to display the name of the sender and the reciever based on message sender type and then filling in the doctor info or patient info as necessary.